Renamed cli package to ohiecli the name used was incorrect and was causing import errors
