+++ ohiecli/cmd/features/build-cli.sh
